+++ R/paris2024_palette.R
@@ -9,6 +9,7 @@
 #' interpolates between colors to create a scale of values. If omitted, function
 #' assumes continuous if n > length of palette, and discrete if n < length of
 #' palette.
+#' @param reverse Whether to reverse colour palette (TRUE) or not (FALSE, default).
 #'
 #' @return A vector of colours.
 #' @export
@@ -17,7 +18,8 @@
 #' paris2024_palette("GamesLook", n = 6, type = "continuous")
 #' paris2024_palette("Logo", 2)
 #' paris2024_palette("OlympicRings")
-paris2024_palette <- function(name, n, type = c("discrete", "continuous")) {
+#' paris2024_palette("OlympicRings", reverse = TRUE)
+paris2024_palette <- function(name, n, type = c("discrete", "continuous"), reverse = FALSE) {
   pal <- paris2024_palettes[[name]]
 
   if (is.null(pal)) {
@@ -36,6 +38,10 @@ paris2024_palette <- function(name, n, type = c("discrete", "continuous")) {
 
   if (type == "discrete" && n > length(pal)) {
     stop("Number of requested colors greater than what palette can offer")
+  }
+
+  if (reverse) {
+    pal <- rev(pal)
   }
 
   out <- switch(type,