+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Modify palette GamesLook
 * Add new palettes: Generation2024, Impact2024, TerreDeJeux and Tickets
+* Add option to reverse colour palette
 * Build README with Rmarkdown
 * Monitor test coverage